requirements.txt
requests
datetime
pydantic
pytz
httptools
pushbullet.py==0.9.1
pymongo
pymongo[srv]
pandas
pandas-ta
ta-lib==0.4.17
